+++ lib/json_schema_utils.dart
@@ -8,3 +8,4 @@ export 'src/json_type.dart' show JsonType;
 export 'src/string_json_schema.dart' show StringJsonSchema;
 export 'src/boolean_json_schema.dart' show BooleanJsonSchema;
 export 'src/number_json_schema.dart' show NumberJsonSchema;
+export 'src/null_json_schema.dart' show NullJsonSchema;