+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add a String type only JSON Schema class
 - Add a Boolean type only JSON Schema class
 - Add a Number type only JSON Schema class
+- Add a Null type only JSON Schema class
 
 
 ## 1.0.0-dev.9